refactor: worklets api usage and dirs
