+++ packages/react-native-audio-worklets/common/cpp/audioworklets/AudioWorkletsRunner.cpp
@@ -39,17 +39,7 @@ AudioWorkletsRunner &AudioWorkletsRunner::operator=(AudioWorkletsRunner &&other)
     return *this;
   }
 
-  if (workletInitialized_) {
-    if (auto strongRuntime = weakRuntime_.lock()) {
-      strongRuntime->runSync([this](jsi::Runtime & /*rt*/) -> jsi::Value {
-        reinterpret_cast<jsi::Function *>(unsafeWorklet_.data())->~Function();
-        workletInitialized_ = false;
-        return jsi::Value::undefined();
-      });
-    } else {
-      workletInitialized_ = false;
-    }
-  }
+  destroyCachedWorklet();
 
   weakRuntime_ = std::move(other.weakRuntime_);
   unsafeRuntimePtr_ = other.unsafeRuntimePtr_;
@@ -65,30 +55,27 @@ AudioWorkletsRunner &AudioWorkletsRunner::operator=(AudioWorkletsRunner &&other)
 }
 
 AudioWorkletsRunner::~AudioWorkletsRunner() {
+  destroyCachedWorklet();
+}
+
+void AudioWorkletsRunner::destroyCachedWorklet() {
   if (!workletInitialized_) {
     return;
   }
 
   auto strongRuntime = weakRuntime_.lock();
   if (strongRuntime == nullptr) {
+    workletInitialized_ = false;
+    unsafeRuntimePtr_ = nullptr;
     return;
   }
 
   strongRuntime->runSync([this](jsi::Runtime & /*rt*/) -> jsi::Value {
     reinterpret_cast<jsi::Function *>(unsafeWorklet_.data())->~Function();
     workletInitialized_ = false;
+    unsafeRuntimePtr_ = nullptr;
     return jsi::Value::undefined();
   });
-}
-
-std::optional<jsi::Value> AudioWorkletsRunner::executeOnRuntimeSync(
-    const std::function<jsi::Value(jsi::Runtime &)> &&job) const noexcept(noexcept(job)) {
-  auto strongRuntime = weakRuntime_.lock();
-  if (strongRuntime == nullptr) {
-    return std::nullopt;
-  }
-
-  return strongRuntime->runSync(job);
 }
 
 std::shared_ptr<AudioChannelViews> AudioWorkletsRunner::createChannelViews(
